requirements.txt
selenium
openai
python-dotenv
requests
streamlit
PyPDF2